rsx: Restructure and simplify some header include chains
