+++ rpcs3/Emu/RSX/GL/GLRenderTargets.cpp
@@ -1,5 +1,6 @@
 #include "stdafx.h"
 #include "GLGSRender.h"
+#include "Emu/RSX/rsx_methods.h"
 
 color_format rsx::internals::surface_color_format_to_gl(rsx::surface_color_format color_format)
 {
