+++ rpcs3/Emu/RSX/VK/VKGSRender.cpp
@@ -6,6 +6,7 @@
 #include "VKRenderPass.h"
 #include "VKResourceManager.h"
 #include "VKCommandStream.h"
+#include "Emu/RSX/rsx_methods.h"
 #include "Emu/Memory/vm_locking.h"
 
 namespace vk
