+++ rpcs3/Emu/RSX/VK/VKVertexBuffers.cpp
@@ -1,7 +1,7 @@
 #include "stdafx.h"
 #include "VKGSRender.h"
-#include "../rsx_methods.h"
 #include "../Common/BufferUtils.h"
+#include "../rsx_methods.h"
 
 namespace vk
 {
